+++ Frameworks/File_Extractor/File_Extractor/unrar/getbits.hpp
@@ -19,7 +19,7 @@ public:
 	{
 		InAddr=InBit=0;
 	}
-	void addbits(long Bits)
+	void addbits(int Bits)
 	{
 		Bits+=InBit;
 		InAddr+=Bits>>3;
@@ -27,13 +27,13 @@ public:
 	}
 	unsigned int getbits()
 	{
-		uint BitField=(uint)InBuf[InAddr] << 16;
+		unsigned int BitField=(uint)InBuf[InAddr] << 16;
 		BitField|=(uint)InBuf[InAddr+1] << 8;
 		BitField|=(uint)InBuf[InAddr+2];
 		BitField >>= (8-InBit);
 		return(BitField & 0xffff);
 	}
-	void faddbits(long Bits);
+	void faddbits(int Bits);
 	unsigned int fgetbits();
 	bool Overflow(int IncPtr) {return(InAddr+IncPtr>=MAX_SIZE);}
 };
